fix anti-windup

+++ Core/Inc/control/pid.h
@@ -27,7 +27,7 @@ typedef struct
     IIRfilter_t error_difference;
 } PID_t;
 
-float PID_update(PID_t *self, float error, int16_t control_difference);
+float PID_update(PID_t *self, float error, float control_difference);
 void PID_update_with_error_injection(PID_t *self, float error);
 void PID_reset(PID_t *self);
